[#177] refactor: repository - repositoryImpl 1대1 대응되도록 수정
