+++ lib/data/datasources/auth_google_datasource_impl.dart
@@ -1,11 +1,16 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 import 'package:wehavit/common/constants.dart';
 import 'package:wehavit/common/constants/firebase_field_name.dart';
 import 'package:wehavit/common/utils/firebase_collection_name.dart';
 import 'package:wehavit/data/datasources/auth_google_datasource.dart';
 import 'package:wehavit/presentation/auth/data/entities/auth_result.dart';
+
+final googleAuthDatasourceProvider = Provider<AuthGoogleDatasource>((ref) {
+  return AuthGoogleDatasourceImpl();
+});
 
 class AuthGoogleDatasourceImpl implements AuthGoogleDatasource {
   @override
