+++ lib/data/datasources/wehavit_datasource.dart
@@ -1,8 +1,14 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:wehavit/common/utils/custom_types.dart';
+import 'package:wehavit/data/datasources/firebase_datasource_impl.dart';
 import 'package:wehavit/domain/entities/confirm_post_entity/confirm_post_entity.dart';
 import 'package:wehavit/domain/entities/reaction_entity/reaction_entity.dart';
 import 'package:wehavit/domain/entities/resolution_entity/resolution_entity.dart';
 import 'package:wehavit/domain/entities/user_data_entity/user_data_entity.dart';
+
+final wehavitDatasourceProvider = Provider<WehavitDatasource>((ref) {
+  return FirebaseDatasourceImpl();
+});
 
 abstract class WehavitDatasource {
   // Friend Repository
