+++ lib/data/models/auth_result_model.dart
@@ -1,0 +1,1 @@
+enum AuthResult { none, aborted, success, failure }
