+++ lib/domain/usecases/login/email_and_password_login_usecase_provider.dart
@@ -1,4 +1,6 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:wehavit/data/repositories/auth_repository_impl.dart';
+import 'package:wehavit/domain/usecases/login/email_and_password_login_usecase.dart';
 import 'package:wehavit/presentation/auth/auth.dart';
 
 final emailAndPasswordLogInUseCaseProvider =
