+++ lib/domain/usecases/login/email_and_password_register_usecase.dart
@@ -1,6 +1,6 @@
 import 'package:wehavit/common/common.dart';
-import 'package:wehavit/presentation/auth/auth.dart';
-import 'package:wehavit/presentation/auth/data/entities/auth_result.dart';
+import 'package:wehavit/data/models/auth_result_model.dart';
+import 'package:wehavit/domain/repositories/auth_repository.dart';
 
 class EmailAndPasswordRegisterUseCase {
   EmailAndPasswordRegisterUseCase(this._authRepository);
