+++ lib/domain/usecases/login/google_login_usecase.dart
@@ -1,6 +1,6 @@
 import 'package:wehavit/common/common.dart';
-import 'package:wehavit/presentation/auth/auth.dart';
-import 'package:wehavit/presentation/auth/data/entities/auth_result.dart';
+import 'package:wehavit/data/models/auth_result_model.dart';
+import 'package:wehavit/domain/repositories/auth_repository.dart';
 
 class GoogleLogInUseCase implements FutureUseCase<AuthResult, NoParams> {
   GoogleLogInUseCase(this._authRepository);
