--- lib/presentation/auth/domain/domain.dart
@@ -1,2 +1,0 @@
-export 'repositories/repositories.dart';
-export 'usecases/usecases.dart';
