--- lib/presentation/auth/domain/repositories/repositories.dart
@@ -1,2 +1,0 @@
-export 'auth_repository.dart';
-export 'auth_repository_provider.dart';
