+++ lib/presentation/auth/presentation/providers/auth_notifier.dart
@@ -1,8 +1,11 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:wehavit/common/utils/no_params.dart';
+import 'package:wehavit/data/models/auth_result_model.dart';
+import 'package:wehavit/domain/usecases/login/email_and_password_login_usecase.dart';
+import 'package:wehavit/domain/usecases/login/email_and_password_register_usecase.dart';
+import 'package:wehavit/domain/usecases/login/google_login_usecase.dart';
+import 'package:wehavit/domain/usecases/logout/logout_usecase.dart';
 import 'package:wehavit/presentation/auth/auth.dart';
-import 'package:wehavit/presentation/auth/data/entities/auth_result.dart';
-import 'package:wehavit/presentation/auth/domain/usecases/login/email_and_password_register_usecase.dart';
 
 class AuthNotifier extends StateNotifier<AuthState> {
   AuthNotifier(
