+++ lib/presentation/auth/presentation/providers/auth_provider.dart
@@ -1,6 +1,9 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:wehavit/domain/usecases/login/email_and_password_login_usecase_provider.dart';
+import 'package:wehavit/domain/usecases/login/email_and_password_register_usecase_provider.dart';
+import 'package:wehavit/domain/usecases/login/google_login_usecase_provider.dart';
+import 'package:wehavit/domain/usecases/logout/logout_usecase_provider.dart';
 import 'package:wehavit/presentation/auth/auth.dart';
-import 'package:wehavit/presentation/auth/domain/usecases/login/email_and_password_register_usecase_provider.dart';
 
 final authProvider = StateNotifierProvider<AuthNotifier, AuthState>((ref) {
   final usecaseGoogleLogIn = ref.watch(googleLogInUseCaseProvider);
