+++ lib/presentation/auth/presentation/providers/auth_state_changes/auth_state_changes_provider.dart
@@ -1,6 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:wehavit/presentation/auth/auth.dart';
+import 'package:wehavit/data/repositories/auth_repository_impl.dart';
+import 'package:wehavit/domain/usecases/auth_state_changes/auth_state_changes_usecase.dart';
 
 final authStateChangesProvider = StreamProvider<User?>((ref) {
   final repository = ref.watch(authRepositoryProvider);
